Swap aliases removed in the latest numpy release.
Summary:
Numpy removed support for np.object, np.bool, np.float, np.complex,
np.str, and np.int with 1.24.0 release. Seems like upstream tensorflow
just replaced these with an underscored version of these types, doing
the same here.

TF1.15 Only

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Maniphest Tasks: T72490

Differential Revision: https://phabricator.sourcevertex.net/D79436

+++ tensorflow/contrib/distributions/python/ops/vector_diffeomixture.py
@@ -1058,7 +1058,7 @@ def softmax(x, axis, name=None):
     axis = ops.convert_to_tensor(axis, dtype=dtypes.int32, name="axis")
     axis_ = tensor_util.constant_value(axis)
     if axis_ is not None:
-      axis = np.int(ndims + axis_ if axis_ < 0 else axis_)
+      axis = np.int_(ndims + axis_ if axis_ < 0 else axis_)
     else:
       axis = array_ops.where_v2(axis < 0, ndims + axis, axis)
   return nn_ops.softmax(x, axis=axis)
